os: alias FileInfo/DirEntry to io/fs (Go-faithful unification)
Go's `os` package does not define its own FileInfo / DirEntry /
FileMode — `os.FileInfo`, `os.DirEntry`, `os.FileMode` are exact type
aliases for the `io/fs` interfaces. goish had diverged: `os` carried a
standalone `pub trait FileInfo` and a concrete `pub struct DirEntry`.

This refactor makes `io/fs` the single canonical home:

  - `io/fs::FileInfo::Size` now returns `int` (goish's Go-facing 64-bit
    int) instead of `i64`, per AGENTS.md §3.
  - `os::FileInfo` / `os::DirEntry` are now `pub use crate::io::fs::*`
    re-exports of the `#[goish::interface]` traits.
  - `os::FileInfoData` keeps its concrete shape and now implements
    `io::fs::FileInfo` with the interface boilerplate
    (`__goish_as_dyn_any` override + lazy registry registration).
  - The old concrete `os::DirEntry` struct becomes the private
    `os::unixDirent`, implementing `io::fs::DirEntry`. It carries its
    parent directory path so `Info()` can `lstat` the entry on demand,
    matching Go's `unixDirent.Info`.
  - `os::ReadDir` now returns a slice of the interface —
    `(slice<Arc<dyn io::fs::DirEntry + Send + Sync>>, error)` — exactly
    mirroring `io/fs::ReadDir`.

Consumers updated: `path/filepath::WalkDir` (closure now takes an
interface-borrow `&dyn os::DirEntry`, root entry synthesised via a
private `filepathDirEnt`). `net/http/fs.rs` needed no change.
`os_readdir_smoke` gains a `DirEntry.Info()` lstat check and is now
registered in Cargo.toml.

cargo check --lib + cargo build --examples clean; make e2e LOOPS=1
all green (160/160).

+++ examples/io_fs_walkdir_smoke.rs
@@ -78,7 +78,7 @@ impl fs::FileInfo for mapInfo {
     fn Name(&self) -> string {
         self.name.clone()
     }
-    fn Size(&self) -> i64 {
+    fn Size(&self) -> int {
         self.size
     }
     fn Mode(&self) -> fs::FileMode {
